requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2
datasets==2.16.1
huggingface-hub==0.20.2

# Database
duckdb==0.9.2

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Dashboard
streamlit==1.30.0
plotly==5.18.0
altair==5.2.0

# AI/ML
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
scikit-learn==1.4.0
sentence-transformers==2.3.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
pyyaml==6.0.1

# Development
pytest==7.4.4
black==24.1.1
flake8==7.0.0
